backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "playwright>=1.44.0",
    "beautifulsoup4>=4.12.3",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.7.1",
    "lxml>=5.2.2", # Parser for BeautifulSoup
    "httpx>=0.27.0", # Explicitly add httpx
]
